+++ lldb/source/Utility/CMakeLists.txt
@@ -50,6 +50,7 @@ add_lldb_library(lldbUtility NO_INTERNAL_DEPENDENCIES
   Listener.cpp
   Log.cpp
   NameMatches.cpp
+  Policy.cpp
   ProcessInfo.cpp
   RealpathPrefixes.cpp
   RegisterValue.cpp
